Don't cd around invocations of install-man
install-man is a config.status-generated script, so Autoconf will
substitute in the path to our install-sh script, but Autoconf prefers
relative paths that are only valid in the directory in which
install-man is generated.  We therefore must avoid changing directories
before calling install-man.

Instead of cd, use relative paths including the directory and adjust
the installation path accordingly.

Change-Id: I8666257b9d74fa80cf7f136bd0ee6ff895897066
Reviewed-on: http://gerrit.openafs.org/2460
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>