Include com_err.h in dumpstuff.c
src/volser/dumpstuff.c uses afs_error_message in its error reporting,
but it does not include com_err.h, so it does not get the prototype
for afs_error_message. The default return type for unprototyped
functions is 'int', so the return value of afs_error_message can get
truncated on platforms where sizeof(char*) > sizeof(int), causing a
segfault whenever the result of afs_error_message is logged.

This is a 1.4-only change. This was fixed in the 1.5 branch in the
midst of the larger warning cleanup effort.

Change-Id: Id3bcde20609130d8e4e393c6c8aae75682bcd966
Reviewed-on: http://gerrit.openafs.org/2526
Tested-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>