Windows: cm_Analyze should treat VOFFLINE like VMOVED or VNOVOL
Volume package bugs in the file server can result in VOFFLINE
being returned to the client instead of VNOVOL or VMOVED.  As
a result the Unix CM treats VOFFLINE the same as VMOVED and VNOVOL.
The Windows client has not.  As a result, bugs in the file server
can cause the Windows client to lose if the volume has in fact
been moved to another server.

As part of this change, the volume location list is updated prior
to the volume status being applied to the server from which the
error was received.

LICENSE MIT

Change-Id: I01036aa9d1999d0ba6822beb1b73500d365bf0b3
Reviewed-on: http://gerrit.openafs.org/2532
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>