Linux: Rework statfs super block operations configure test
The configure test to detect if the statfs super block operation
needs a dentry argument is based on vfs_statfs, and assumes that
its signature matches the one of the operation.  In 2.6.36 this is
no longer true and the test fails.

Rework the test to actually test the operation we're interested in.

This change is required for 2.6.36.

Reviewed-on: http://gerrit.openafs.org/2545
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>

(cherry-picked from commit 08552ad8a94f7cc5908aabe8385711e09a6418e4)

Reviewed-on: http://gerrit.openafs.org/2553
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit b8b708fc416aa281fe1d1ce1d498cc381c81679c)
Change-Id: Ib40de2018ecec5e3c82d8fa829364c31d29526b4
Reviewed-on: http://gerrit.openafs.org/3516
Reviewed-by: Simon Wilkinson <sxw@inf.ed.ac.uk>