Fixes for signed char architectures
"char" is unsigned on s390x, which causes a few warnings where we
test for negative values.

Add a signed modifier in a few cases, and in vos.c replace with an
int to match a similar section of code in the same file.

Mostly a warning fix, but in one case the variable could be used
to return a -1 error value, so it may have caused an actual bug.

Change-Id: I005fbcd7c573735df7403316ec0ced67ee81e559
Reviewed-on: http://gerrit.openafs.org/2602
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>