FBSD CM: don't call afs_close when recycling
Don't call afs_close when handling VOP_CLOSE on a recycled
vnode, since there was no matching open.  This corrects the
opens count, which was seen to go have gone negative in the
reclaim vop.  For clarity, assert if afs_vop_close is entered
with a VI_DOOMED vnode and avc->opens != 0.

Change-Id: I511a4f2a924c2f8e20f3ecdaa445fbe803289a47
Change-Id: I1b2307fd3318fa54e8f7fb72a5d3f843e2a38404
Reviewed-on: http://gerrit.openafs.org/2612
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>