Update the man pages to discourage use of uss
Usually the text added was a copy of a CAUTION section that
had already been added in a few places. This change applies it
consistently across all uss-related man pages. In pod1/afs.pod that
text would be excessively wordy; a briefer note is used there to
direct the reader to the full text.

This is a partial fix for RT bug #128018. It does not fully close
the bug; the AFS Administrators Guide needs to be updated as well.

Reviewed-on: http://gerrit.openafs.org/2637
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
(cherry picked from commit 61ddfbaa19f28bd8b352dea307ab752239470026)

Change-Id: I18596dd78d0d5875d1aa442098c6bd757f496953
[sxw@your-file-system.com fixed the commit message for this pullup]
Reviewed-on: http://gerrit.openafs.org/4011
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>