vol: Set but not used variables
Fix some set but unused variable warnings in the vol directory.

In VDetachVolume_r, much of the code and variables are only
useful in the FSSYNC_BUILD_CLIENT case.  Scope the code
accordingly.

forcefree is only used in the AFS_DEMAND_ATTACH_FS case, so
scope code that sets it accordingly.

Spotted by gcc 4.6

Change-Id: I345db6c9aba2d433ccbbc33fdc10d716c06b913c
Reviewed-on: http://gerrit.openafs.org/2786
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>